object 유틸 함수 파일 통합
