+++ src/utils/content/postMetadata.ts
@@ -8,7 +8,7 @@ import { Locale } from '@/constants/i18n';
 import { allTranslationNumber } from '@/constants/stats';
 import { sliceByPage, sortByDate } from '@/utils/core/array';
 import { slugify } from '@/utils/core/format';
-import { pick } from '@/utils/core/pick';
+import { pick } from '@/utils/core/object';
 
 const sortedPostMetadata = sortByDate(postMetadata);
 const sortedEnPostMetadata = sortByDate(enPostMetadata);
